+++ tests/e2e/mocks/cache.ts
@@ -1,0 +1,46 @@
+export const mockedCacheEntries = [
+  [
+    'SGVhbHRoMQ==e___R28gZm9yIGEgd2Fsay==',
+    {
+      id: 'R28gZm9yIGEgd2Fsay==',
+      name: 'Go for a walk',
+      scale: 'Atom',
+      description: 'A daily walk to improve cardiovascular health.',
+      startTime: 1617235200,
+      endTime: 1617321600
+    }
+  ],
+  [
+    'SGVhbHRoMQ==e___TGlmdCB3ZWlnaHRz',
+    {
+      id: 'TGlmdCB3ZWlnaHRz',
+      name: 'Lift weights',
+      scale: 'Atom',
+      description: 'Strength training to build muscle and increase metabolism.',
+      startTime: 1617235200,
+      endTime: 1617321600
+    }
+  ],
+  [
+    'SGVhbHRoMQ==e___TWFrZSBhIGhlYWx0aHkgbWVhbA==',
+    {
+      id: 'TWFrZSBhIGhlYWx0aHkgbWVhbA==',
+      name: 'Make a healthy meal',
+      scale: 'Atom',
+      description: 'Preparing nutritious meals to fuel the body for optimal health.',
+      startTime: 1617235200,
+      endTime: 1617321600
+    }
+  ],
+  [
+    'SGVhbHRoMQ==e___UHJhY3RpY2UgeW9nYQ==',
+    {
+      id: 'UHJhY3RpY2UgeW9nYQ==',
+      name: 'Practice yoga',
+      scale: 'Atom',
+      description: 'Engage in yoga to enhance flexibility, strength, and mental clarity.',
+      startTime: 1617235200,
+      endTime: 1617321600
+    }
+  ]
+]
