+++ tests/setupTests.ts
@@ -55,28 +55,3 @@ jest.mock('d3-zoom', () => ({
     // Mock other D3 methods as needed
   }),
 }));
-
-// Mock atoms for setting hierarchy bounds
-const mockBreadthBoundsAtom = atom({});
-const mockDepthBoundsAtom = atom({});
-const mockCurrentSphereAtom: PrimitiveAtom<SphereHashes> = atom({ entryHash: SPHERE_ID, actionHash: "" });
-
-// Mock setters for hierarchy bounds
-const mockSetBreadthBounds = jest.fn();
-const mockSetDepthBounds = jest.fn();
-
-jest.mock('../app/src/state/currentSphereHierarchyAtom.ts', () => ({
-  setBreadths: (sphereHash: string, bounds: number[]) => mockSetBreadthBounds(sphereHash, bounds),
-  setDepths: (sphereHash: string, bounds: number[]) => mockSetDepthBounds(sphereHash, bounds),
-  currentSphere: mockCurrentSphereAtom,
-  currentSphereHierarchyBounds: atom((get) => {
-    return {
-      [SPHERE_ID]: {
-        minDepth: 0,
-        maxDepth: 1,
-        minBreadth: 0,
-        maxBreadth: 0,
-      },
-    };
-  }),
-}));
