+++ tests/utils-frontend.tsx
@@ -1,0 +1,14 @@
+
+import { Provider } from "jotai/react";
+import { useHydrateAtoms } from "jotai/utils";
+
+const HydrateAtoms = ({ initialValues, children }) => {
+  useHydrateAtoms(initialValues)
+  return children
+}
+
+export const TestProvider = ({ initialValues, children }) => (
+  <Provider>
+    <HydrateAtoms initialValues={initialValues}>{children}</HydrateAtoms>
+  </Provider>
+)
